+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
-
+## <next>
+* Added storybook
+* Modified `Header.Basic` prop types to allow string content in `left`, `center` and `right` props.
 ## 3.0.2
 * Replace theme files with `@opuscapita/oc-cm-theme` package